Create people.JSON
