+++ .rubocop.yml
@@ -9,6 +9,8 @@ AllCops:
 
 Layout/LineLength:
   Max: 120
+Layout/EndOfLine:
+  Enabled: false
 Metrics/MethodLength:
   Max: 20
 Metrics/AbcSize:
