+++ book.rb
@@ -1,4 +1,5 @@
 require './rental'
+
 class Book
   attr_accessor :title, :author, :rents
