+++ teacher.rb
@@ -1,5 +1,7 @@
 require './person'
 class Teacher < Person
+  attr_reader :specialization
+
   def initialize(specialization, *args, **kwargs)
     super(*args, **kwargs)
     @specialization = specialization
